efw-protocols: add implementation for protocol about port configuration
This commit adds implementation for protocol about port configuration.
The batch of code is moved from internal transaction module to new
module.
