+++ libs/efw/runtime/src/model.rs
@@ -119,7 +119,7 @@ impl CtlModel<hinawa::SndEfw> for EfwModel {
             Ok(true)
         } else if self.input_ctl.read(unit, elem_id, elem_value, TIMEOUT_MS)? {
             Ok(true)
-        } else if self.port_ctl.read(unit, elem_id, elem_value)? {
+        } else if self.port_ctl.read(unit, elem_id, elem_value, TIMEOUT_MS)? {
             Ok(true)
         } else if self.guitar_ctl.read(unit, elem_id, elem_value)? {
             Ok(true)
@@ -145,7 +145,7 @@ impl CtlModel<hinawa::SndEfw> for EfwModel {
             Ok(true)
         } else if self.input_ctl.write(unit, elem_id, old, new, TIMEOUT_MS)? {
             Ok(true)
-        } else if self.port_ctl.write(unit, elem_id, old, new)? {
+        } else if self.port_ctl.write(unit, elem_id, old, new, TIMEOUT_MS)? {
             Ok(true)
         } else if self.guitar_ctl.write(unit, elem_id, old, new)? {
             Ok(true)
